Add tests to `lint-v2` and `fmt-v2` (#8668)
### Problem

We weren't testing `lint`, which has an edge case when one of the targets results in a lint failure. `fmt` had an integration test which can be rewritten to a less costly unit test.

### Result

We test more edge cases.

It should be easier to write future tests for these files and to write `run_rule()`-based tests for other rules, as the more examples we have of how to test rules, the easier it will become.