Bump default Pytest to 4.6 and add warning for future change to Pytest 5 (#8648)
### Problem
We're using an old version of Pytest from 2018 and are two major releases behind. This means users are missing out on new features like Pytest 5.0 enabling `faulthandler` by default and that certain plugins won't work because Pytest is too outdated.

However, Pytest 5.0 drops support for running tests with Python 2.

### Solution
Set the default Pytest to 4.6.x, which is the last minor series to support Python 2. This brings nice benefits to users without being majorly breaking.

Then, we add a deprecation for users not specifying `--pytest-version` to ensure that they don't get bit by us switching to Pytest 5 in two major releases.

Finally, we set our own internal Pytest to the newest release so that we get the newest and greatest, e.g. default `faulthandler`.

### Result
Users who have not specified `--pytest-version` will get this warning:

> [WARN] /Users/eric/DocsLocal/code/projects/pants/src/python/pants/backend/python/tasks/pytest_prep.py:92: DeprecationWarning: DEPRECATED: Pants defaulting to a Python 2-compatible Pytest version will be removed in version 1.25.0.dev2.
  Pants will soon start defaulting to Pytest 5.x, which no longer supports running tests with Python 2. In preparation for this change, you should explicitly set what version of Pytest to use in your `pants.ini` under the section `pytest`.
>
> If you need to keep running tests with Python 2, set `version` to `pytest>=4.6.6,<4.7` (the current default). If you don't have any tests with Python 2 and want the newest Pytest, set `version` to `pytest>=5.2.4`.