Fix ant compatible xml reporting for scala tests
Recent changes to ScalaTestUtil (specifically the ScalaTestJunitRunnerWrapper facade) broke the xml reporting (AntJUnitXmlReporter) because the description of the runner had changed and didn't describe children of the suite correctly.

The solution implenented calls the getDescription method on the delegate
as should be.

This fixes the regression introduced by replacing the original
JUnitRunner of scalatests by ScalaTestJunitRunnerWrapper