Rewrite Black tests from `PantsRunIntegrationTest` to rules-style integration tests (#8684)
Traditionally, we'd write integration tests via `PantsRunIntegrationTest`, which has the cost of starting a new Pants process and is also not very isolated/hermetic. Instead, with V2 rules, we can write integration tests using `self.scheduler.product_request()`. This is preferable because it runs faster and better isolates the rule we are testing away from other parts of Pants.

For example, in https://github.com/pantsbuild/pants/pull/8683, we are going to default to Black not being turned on as a plugin within our codebase. That causes the integration tests to fail because they are no longer able to recognize the plugin. While we could work around that via setting up the Pants config within the ITs themself, the more sound solution is to use our `self.scheduler.product_request()`-based tests, which have a similar effect of integration tests but in a more controlled and reusable way.