Move Black into a plugin to enable turning it on and off via `backend_plugins` (#8683)
### Problem

We need some mechanism to turn off autoformatters and linters in V2 other than the current `--fmt-{isort,scalafmt,etc}-skip`, which doesn't work because that uses a task-specific option system.

Ideally, the option will be tool-specific, so you can say something like:

```ini
[black]
enable: True

[mypy]
enable: False
```

However, the solution should also be hygienic. If, for example, someone doesn't want to use Black, then we don't want Black's rules to pollute the rule graph. Nor do we want to make rule authors provide a no-op implementation.


### Solution
We decided in https://github.com/pantsbuild/pants/issues/8346 to turn each linter into its own plugin, meaning it may be enabled via the `backend_packages` option like this:

```ini
[GLOBAL]
backend_packages: +[
    "pants.backend.python.lint.black",
    "pants.backend.python.lint.isort",
    "pants.backend.scala.lint.scalafmt",
  ]
```

When specified, that plugin will be enabled, meaning that its options may be configured and its rules will be automatically registered against `fmt` and `lint`.

### Result

We have a scalable and explicit mechanism for users to enable formatters and linters in a way that also makes it effortless for rule authors to have a way to opt-out of their tool.

This only implements Black. In a followup, we will implement isort under the plugin `pants.backend.python.lint.isort`.

For now, V1 tasks will keep the same mechanism as before. We could move them towards this new plugin system, but save that for a followup if we choose to make the deprecations necessary to move V1 users towards the new system.