Treat all `MaterializeToDirectories` paths as relative to the buildroot (#8696)
### Problem

We want to be able to safely use relative paths when writing rules, such as `binary.py` using the path `dist/` and `fmt-v2` uses the path `./`. Currently, Rust will accept this as-is, thus creating the directory relative to the cwd, instead of relative to the build root.

There is no use case we can imagine where we want to allow writing to an absolute path. For testing, our `TestBase` code already sets up a new temporary build root for every test. When we need a temporary directory, we can set that to live within `.pants.d`.

### Solution

Modify the Rust code to join the final destination path with the build root.

Also, add a runtime Python check that every passed path is relative. Otherwise, the engine would fail.

Finally, fix an oversight in `TestBase` so that `_init_engine()` passes the temporary `build_root` property to the engine session creation. Otherwise, the session will use the global `get_buildroot()`, which we want to avoid due to persisting state between tests.