Use default argument for `DirectoryToMaterialize`'s `path_prefix` (#8711)
### Problem

We now always materialize directories relative to the build root (https://github.com/pantsbuild/pants/pull/8696). In making that change, the vast majority of use cases showed that we usually want to write at the root level of the build root.

For example, if the digest is for a file called `tests/foo.txt` and a file called `bar.txt`, then we would want the build root to end up with the files `/User/.../pants/bar.txt` and `/User/.../pants/tests/foo.txt`.

Right now, all of those call sites must set `path=""`, which results in a lot of boilerplate and is also not very intuitive what that means. 

### Solution

Use a default argument, which we could not do until recently now that we have dataclasses.

This requires reordering the args for the dataclass, which results in a small Rust change.