add types to await Get[X](...)!! (#8697)
### Problem

While #8639 allowed mypy type-checking of `@rule`s, it didn't do quite as much as it could have. In particular, while we already require explicitly writing the return type in every `Get(X, ...)` call, we don't yet have mypy automatically inferring the return type of the `await Get(X, ...)` must be `X`.

### Proposed Solution

- Have `Get` and `MultiGet` both mix in `Generic[_Product]`, and use the inferred `_Product` type to deduce the `ReturnType` for the `__await__` generator.
- Override `__class_getitem__` on `Get` to allow `Get[X](...)` to be converted to `Get(X, ...)`.
- Modify `rules.py` to recognize both `Get(...)` and `Get[X](...)` calls in the `_RuleVisitor`.

#### Other Solutions Considered

1. An attempt was made to parameterize the `Get`'s *`subject_declared_type`* and make that available on the type level as well:
```python
@rule
async def f() -> Y:
  # In this PR now (s: Snapshot is inferred):
  s = await Get[Snapshot](Digest(...))
  # What we might like to try (s: Snapshot is inferred,
  # d: Digest is also inferred and type-checked with mypy!):
  s = await Get[Snapshot, Digest](d)
  ...
```

This wasn't implemented yet because it wasn't as clear that that was as useful as just the product type, and it seemed better to put that into a followup PR. A remaining question is whether/how we can integrate the existing 2-argument `Get` constructor syntax while also inferring types as above.

2. It might be possible to avoid having to explicitly type out the square brackets for `Get[X](...)`, and instead try to infer that from the existing constructor arguments (i.e. it might be possible to have the appropriate mypy types without changing the `Get(X, ...)` calls that exist now).

Since the return type of a `Get` right now is always a concrete type (not an `@union` or something), it seemed like we didn't lose much information by making it into a type argument instead, and it seemed nicer to explicitly annotate the type to make it easier to quickly scan code and understand what the output type of a `Get[X]` is supposed to be.

### Result

`await Get[X](...)` calls will now have mypy infer the return type `X`!