Refactor `fmt-v2` to use Workspace (#8691)
### Problem

When writing the original `fmt-v2` goal, `Workspace` was not yet landed in master. So, we had to hackily use Python file system APIs to save the `FmtResult`'s digest into the actual workspace. This works but is not ideal because it does not use V2 abstractions and has a performance hit— materializing `FilesContent` to memory is expensive and the old implementation missed out on parallelism.

### Solution

Use `Workspace` and add unit tests for it.

This also addresses the concern from https://github.com/pantsbuild/pants/pull/8684#pullrequestreview-321465344 that we should be testing how `fmt-v2` and `lint-v2` operate against multiple targets.

### Result

* More idiomatic code
* More comprehensive tests
* Detection of conflicting changes, compared to silently overriding each other
* Improved performance