[IntelliJ] Export source jars of dependencies in `export-dep-as-jar` (#8723)
### Problem

In `export-dep-as-jar`, the only thing we export for each dependency is it's compiled jar, `z.jar`. This means that when a user tries to `Go To Definition` of a class defined in a source dependency, it will direct to the (decompiled) classfile:

<img width="1070" alt="Screenshot 2019-11-27 at 14 44 29" src="https://user-images.githubusercontent.com/5861182/69733063-a7353b80-1124-11ea-9a14-91a22d57c6de.png">

### Solution

Bundle up the sources field of dependencies into a `-sources.jar`, and add that to the export of the project via the `sources` Maven classifier. This feature is hidden behind a flag, because it might affect performance.

### Result

When the feature is enabled in the IntelliJ plugin, the dependencies will be imported with sources attached to them:

<img width="1408" alt="Screenshot 2019-11-27 at 14 49 06" src="https://user-images.githubusercontent.com/5861182/69733349-1ca10c00-1125-11ea-84ee-becbc0255323.png">