Refactor V2 Black and isort for better de-duplication (#8732)
### Problem

It was proposed in https://github.com/pantsbuild/pants/pull/8729#discussion_r351581286 that `BlackSetup` and `IsortSetup` were not doing the right thing by including information about the target in the dataclasses. Instead, those should only have the setup that is used for every target, such as the config file location.

### Solution

Following the above recommendation leads to substantial de-duplication, particularly in the implementation of `fmt` and `lint`. New `BlackArgs` and `IsortArgs` dataclasses are created to capture the main difference between `fmt` and `lint` of whether they should pass `--check` in their CLI args. This allows creating new rules to return the `ExecuteProcessResult`.

### Result

Fewer lines of code and improved readability. Better caching because now `BlackSetup` and `IsortSetup` can be used across many targets.