No-op when V2 `fmt`, `lint`, and `test` have a target with empty sources (#8730)
### Problem

`fmt-v2` and `lint-v2` currently check if the `sources` field is not defined and no-ops if so, but it fails to handle the case where `sources` is defined but empty:

```python
python_library(
  name="empty",
  sources=[],
)
```

This pattern appears in our codebase because we sometimes set up targets like `src/python/pants/backend/python/lint/black:tests` that do not yet have any test targets.

For isort, this causes the tool to try to run against all of the 3rd party dependency code included in the PEX, which we do not want to do. Instead, we should no-op.

### Solution

Tweak the logic in `fmt.py` to not only check if `sources` is defined, but also if it's empty.

The Python test runner already had logic to handle this case, but this PR moves that logic one level of abstraction higher from `python_test_runner.py` to `test.py`. Now, we get this no-op behavior for any future test runners we add.

Finally, this adds tests in `rules/core` for all these cases, which involves refactoring `test_test.py`.

#### Possible follow up: abstract this logic even higher

It would be nice if `test`, `fmt`, and `lint` could be passed targets that are already known to have sources so that this logic could be factored out. Presumably, other rules like `repl` and `run` would also only want to work with targets with sources. It's not clear where this logic should go.