Allow V2 isort to use multiple config files (#8731)
### Problem

Isort allows using multiple config files, unlike other tools like Black. We rely on this fact in the Pants repo for formatting of `contrib` vs `examples` vs. everything else. This is necessary so that `contrib` correctly treats `pants.contrib` as first-party but `pants` and `pants_test` as 3rd party.

V2 isort needs to support this reality for us to use it in the Pants repo.

### Solution

Change the `type` of `--config` to `list`.

### Result

Users with a single config file can still use it the same way as before, thanks to how we parse list options: 

```bash
./pants --isort-config=".isort.cfg" fmt-v2 ::
```

Users with multiple config files can specify them as a list:

```ini
[isort]
config: [".isort.cfg", "contrib/.isort.cfg"]
```