Remove deprecated option mypy_version (#8751)
### Problem

The version bumps for the 1.24.devX release cycle triggered the deprecation warnings of option `mypy_version` of the mypy plugin. It seems that this option has reached its end of life.
Relevant slack thread: https://pantsbuild.slack.com/archives/C15HUUSL9/p1575369599013500

### Solution

Remove the options and its usages in the pants codebase.

### Result

Plugin consumers will have to switch to using the `--version` option of the plugin.
The release cycles are unblocked.