Include transitive target dependencies in export-dep-as-jar
Previously, `export-dep-as-jar` did not include the full list of
transitive target dependencies. This meant that client tools like the
IntelliJ compiler and Bloop export were left to traverse the build graph
in order to construct the full classpath of a target.

Now, the output of `export-dep-as-jar` includes a new field
`transitive_targets` field that is similar to `targets` except that it
includes transitive target dependencies respecting `strict_deps`.  For
example, assuming `c -> b -> a` the value of `transitive_targets` is the
following:

```diff
++ new export-dep-as-jar output
 {
   "c:c": {
     "targets": ["b:b"],
+    "transitive_targets": ["b:b", "a:a"],
   }
 }
```
This change allows clients to faithfully reproduce the same classpath as
`./pants compile` without exposing `strict_deps` into the
`export-dep-as-jar` output.