Fix V2 binary to work with multiple targets (#8781)
`./pants --no-v1 --v2 binary target1 target2` would fail with the exception message:

```
 ValueError: dist/ appeared more than once. All paths must be unique.
```

This is because the results are all trying to materialize directory digests with the path prefix `dist/`, which is not allowed. Instead, we can merge them into one digest and materialize the single digest (this is the same pattern we use for `fmt.py`).