Fix pantsbuild wheel publishing.
The `pantsbuild.pants.testinfra` distribution has been broken since
teh `1.23.0.dev1` release with an invalid descrition field. This has
prevented publishing of the package as well as any packages following
it in the release. Fixup the root problem and add a check that will
fail PR CI in the future for similar breaks.

In addition, cleanup a warning against our packages not declaring
`long_description_content_type` brought to light by the new check.