[V2 UI] Pipe Python-side `Console` to the engine when --v2-ui is set (#8771)
### Problem

This commit is a first chunk of work necessary for fixing #8062 . When the --v2-ui flag is enabled, we currently activate the Rust EngineDisplay system which grabs control of the terminal and sets it to raw mode, while simultaneously writing to stdout with the Python Console. This is bad and needs to be fixed.

We eventually want to solve this by having EngineDisplay be fully in charge of any text that needs to get written to the terminal while pants is running. The first step for doing this is to hook up Console to EngineDisplay.

### Solution

I'm going with the third solution @blorente mentioned in his comment on the above issue. When Console is initialized, if --v2-ui is set, then we pass it a reference to the current Python scheduler session, which in turns holds a reference to a Rust Session, which is the data structure that contains an EngineDisplay. When Console invokes one of its methods that write to stdout or stderr, it will pass the string to be printed plus a reference to the Session over FFI to two new functions write_stdout and write_stderr. These in turn will call similar methods on Session.

As of this commit, these methods on Session will simply write to stdout/stderr from Rust instead of from Python, but in a future commit it will be easy to hook these up to EngineDisplay and correctly handle printing this text to the console.

### Result

No user-visible changes should result from this commit.