Load the bootstrapped zinc compiler from the zinc server's classpath. (#8753)
This PR is a dependee of #8750. View that PR for a full description of the problem this change is solving.

Our current zinc wrapper does not implement zinc's scala compiler cache, and instead creates a scala compiler with a ClassLoaderCache which may contain the previously loaded scala compiler class files. This change sets the scala compiler class paths to their locations in the nailgun server directory, so that successive scala compilers hit the class loader cache when being instantiated with different working directories with nailMain.