Fix V2 linters needing to implement a formatter no-op (#8795)
### Problem

Our old `fmt/lint` implementation assumed that every formatter is also a linter and vice versa. It is safe to assume that every formatter is a linter, but it is not safe to assume that every linter is a formatter. For example, Flake8 and Pylint are solely linters.

Specifically, linters had to implement a rule that went from `PythonFormattableTarget -> FmtResult` or the engine would fail with:

```
Exception message: Rules with errors: 1
  (Fmt, [Console, HydratedTargets, Workspace, UnionMembership], [Get(FmtResult, PythonTargetAdaptor), Get(FmtResult, PythonAppAdaptor), Get(FmtResult, PythonBinaryAdaptor), Get(FmtResult, PythonTestsAdaptor), Get(FmtResult, PantsPluginAdaptor), Get(Digest, DirectoriesToMerge)], fmt()):
    No rule was available to compute Get(FmtResult, PantsPluginAdaptor) with parameter types (Address+BuildFileAddress+BuildFileAddresses+Console+Digest+DirectoriesToMerge+DirectoryWithPrefixToAdd+DirectoryWithPrefixToStrip+ExecuteProcessRequest+InputFilesContent+InteractiveRunner+LLVM+MultiPlatformExecuteProcessRequest+NativeToolchain+OptionsBootstrapper+OwnersRequest+PantsPluginAdaptor+PathGlobs+Scope+Specs+ToolchainVariantRequest+UrlToFetch+Workspace)
    No rule was available to compute Get(FmtResult, PythonAppAdaptor) with parameter types (Address+BuildFileAddress+BuildFileAddresses+Console+Digest+DirectoriesToMerge+DirectoryWithPrefixToAdd+DirectoryWithPrefixToStrip+ExecuteProcessRequest+InputFilesContent+InteractiveRunner+LLVM+MultiPlatformExecuteProcessRequest+NativeToolchain+OptionsBootstrapper+OwnersRequest+PathGlobs+PythonAppAdaptor+Scope+Specs+ToolchainVariantRequest+UrlToFetch+Workspace)
    No rule was available to compute Get(FmtResult, PythonBinaryAdaptor) with parameter types (Address+BuildFileAddress+BuildFileAddresses+Console+Digest+DirectoriesToMerge+DirectoryWithPrefixToAdd+DirectoryWithPrefixToStrip+ExecuteProcessRequest+InputFilesContent+InteractiveRunner+LLVM+MultiPlatformExecuteProcessRequest+NativeToolchain+OptionsBootstrapper+OwnersRequest+PathGlobs+PythonBinaryAdaptor+Scope+Specs+ToolchainVariantRequest+UrlToFetch+Workspace)
    No rule was available to compute Get(FmtResult, PythonTargetAdaptor) with parameter types (Address+BuildFileAddress+BuildFileAddresses+Console+Digest+DirectoriesToMerge+DirectoryWithPrefixToAdd+DirectoryWithPrefixToStrip+ExecuteProcessRequest+InputFilesContent+InteractiveRunner+LLVM+MultiPlatformExecuteProcessRequest+NativeToolchain+OptionsBootstrapper+OwnersRequest+PathGlobs+PythonTargetAdaptor+Scope+Specs+ToolchainVariantRequest+UrlToFetch+Workspace)
    No rule was available to compute Get(FmtResult, PythonTestsAdaptor) with parameter types (Address+BuildFileAddress+BuildFileAddresses+Console+Digest+DirectoriesToMerge+DirectoryWithPrefixToAdd+DirectoryWithPrefixToStrip+ExecuteProcessRequest+InputFilesContent+InteractiveRunner+LLVM+MultiPlatformExecuteProcessRequest+NativeToolchain+OptionsBootstrapper+OwnersRequest+PathGlobs+PythonTestsAdaptor+Scope+Specs+ToolchainVariantRequest+UrlToFetch+Workspace)
```

### Solution

Replace `TargetWithSources` with `FormatTarget` and `LintTarget`. Beyond clearer naming, this decouples the `fmt` and `lint` rule so that one may be implemented without the other needing to also be implemented.

We keep `FormatPythonTarget` as satisfying the union for both `FormatTarget` and `LintTarget` so that Black and Isort continue to work without any changes. For those two rules, we assume that if something is formattable, it's also lintable.

However, we introduce a new `LintPythonTarget` that _only_ satisfies the union for `LintTarget`, which allows us to remove the `fmt` rule from Flake8.

### Result

We can now implement traditional linters without needing to also implement format logic.

Down the road, if we ever encounter a tool that only provides formatting but not also linting (i.e. no `--check` flag), then we will be able to extend this setup to get that working too.