Rework `Goal.Options` so that V2 goals work with MyPy (#8742)
### Problem

MyPy does not understand `Goal.Options`, e.g. `List.Options` and `Fmt.Options`, because the implementation relies on runtime creation of the class. MyPy only can understand classes declared at compile time. This is the same reason we needed to rework `pants.util.objects.enum`, `pants.util.objects.datatype`, and `pants.engine.objects.Collection`.

At the same time, we want to preserve as much of the original API as possible because it provided a useful coupling between the `Goal` class and its underlying subsystem. We need to preserve features like the goal's docstring populating `./pants help`, for example.

One of the challenges is that the original implementation requires the Subsystem inner class to have knowledge of the Goal outer class. This is not a well-supported idiom in Python: https://stackoverflow.com/questions/2024566/access-outer-class-from-inner-class-in-python.

### Attempted solutions

#### Attempt #1: pass outer Goal directly to inner Options

I started by trying to implement https://stackoverflow.com/a/7034206. The issue is that we make heavy use of `@classmethod` and `@classproperty`, so we need to be able to access `outer_cls` at the `cls` scope and this pattern only allows using it at the instance scope (`self`).

#### Attempt #2: `Goal.Options` directly instantiated with reference to `Goal`

I then tried this pattern:

```python
class List(Goal):
  name = "list"
  
  @classmethod
  def register_options(cls, register):
    ...

class ListOptions(Options):
  goal_cls = List

@rule 
def list(..., options: ListOptions) -> List:
   ...
```

This required rewriting `rules.py` to ensure that the `Options` were properly registered with the rule graph:

https://github.com/pantsbuild/pants/blob/b88dc0f7d9c0ce322714b74d78f312b4c4b847a2/src/python/pants/engine/rules.py#L138-L155

This worked fine when a rule explicitly requested the `Options` in its rule params, but it meant that rules without explicitly registered options, like `fmt` and `lint`, were no longer registered with the engine! There was no way to feed in the sibling `FmtOptions` or `LintOptions` because these would have to be explicitly created and somehow passed to the engine, even though the corresponding rules had no need to directly consume those subsystems.

### Solution: new `GoalSubsystem` class

Setting up a `Goal` will now involve two steps:
1. Register a `GoalSubsystem` that has the name of the goal, the docstring, and any options (if any).
1. Register a `Goal`, which only needs to point `subsystem_cls` to the sibling `GoalSubsystem`.

For example:

```python
class ListOptions(GoalSubsystem):
  """List all the targets."""
  name = "list"

  @classmethod
  def register_options(cls, register):
    ...

class List(Goal):
  subsystem_cls = ListOptions
```

Rules may then explicitly request the `GoalSubsystem`, e.g. `ListOptions`.

Consumers of the `Goal`, like `rules.py` and `engine_initializer.py`, simply access `Goal.subsystem_cls` to get the attributes they previously accessed directly.

#### Conceptual motivation

This strengthens the concept that Subsystems are the sole mechanism for consuming options in V2. The `GoalSubsystem` may be seen as the external API—containing all the options, the goal name, and the goal description seen by the outside world. Meanwhile, the `Goal` consumes that subsystem (just like the `black.fmt` rule consumes the `Black` subsystem) to produce a new type used internally by the engine.

### Result

MyPy now understands V2 goals, which unblocks us from using MyPy with V2 code and tests (`TestBase` and `PantsRunIntegration` are blocked by this issue).