Rename V2 goals to use the same naming scheme (#8774)
### Problem

When V2 goals conflict with V1 goals so that we are not able to reuse the same name, we have two conventions of naming the V2 goal:

* fast-rule, e.g. `fast-filedeps`
* rule-v2, e.g. `fmt-v2`

It is confusing to use two different conventions. There is no meaningful distinction between the two, only the history of what the rule author chose to use.

Further, both conventions are much longer to type than ideal. We want it to be extremely easy to run the V2 rule through the CLI.

### Solution

Converge on a new convetion of `rule2`. For example, `filedeps` is already claimed by V1 so we name the V2 rule `filedeps2`.