move the subdirectory for exception logs to .pids/ (#8511)
### Problem

Right now, if pants fails with an uncaught exception during import time, it will write a timestamped stacktrace to `<buildroot>/logs/exceptions.log`. This is a little confusing for users, who don't know whether they're supposed to gitignore it or check it in.

### Solution

- Use `.pids` instead of `logs`.
- Fix our `.gitignore`.

### Result

It's now more clear that these logs are specifically generated by pants, and that they can safely be gitignored and/or deleted. Fixes #8809.