Fix terminal display with v2 UI (#8821)
### Problem

The v2 console UI currently does some unintuitive things when modifying the terminal to display the live-updating swim lane display while pants is executing. In tmux, it leaves garbled remnants of the text making up the swim lane display in the tmux scrollback, and in a normal terminal it erases all of the previous terminal scrollback.

### Solution

Modify EngineDisplay to make use of the termion crate's AlternateScreen abstraction when displaying the live swim lanes. This results in automatically saving and restoring the previous terminal state, including scrollback state.

### Result

After the live swim lane display has exited, the terminal scrollback left behind by pants should be sane. This change also seems to remove a bug I was seeing where some log output was being written to the screen when it was in raw mode, resulting in newlines not scrolling back to the left of the screen.