V2 console UI: fix bug with InteractiveProcessRequest (#8833)
### Problem

When running goals that make use of InteractiveProcessRequest such as run, the terminal refreshing would cause any output from the running program to be erased when pants shut down the swim lanes display before exiting.

### Solution

I added some infrastructure to the InteractiveProcessRequest intrinsic that would grab a reference to the EngineDisplay and temporarily disable the interactive console while doing so. This way, any output from a program run with InteractiveProcessRequest would be left in the terminal scrollback like any other normal output.

I also cleaned up a few more small nits around EngineDisplay - namely, displaying the amount of time a node has been running in the swim lanes before rather than after the name of the node (so it doesn't get pushed off the side of the screen), and removing the running bool on that type, since nothing was actually using it.