Add new MyPy subsystem and deprecate `--lint-mypy-config-file` and `--lint-mypy-version` (#8829)
We will soon want to port MyPy to V2. When doing this, we will need to introduce a subsystem for MyPy to take options like `--version` and `--args`. 

We can ease the migration from V1 to V2 by introducing this subsystem now and deprecating the old style of options, even though we do not yet introduce a V2 implementation of MyPy.

However, we only move `--version` and `--config` to the subsystem because it's not yet clear that the V2 implementation should support `--include-requirements` (maybe possible through other mechanisms like `--extra-requirements`) and `--whitelist-tag-name/--verbose`. We can always bring those over to the subsystem later.