jar: default to Force=True
Problem

At present it is impossible to deterministically understand what version
of a dependency is used.  Similarly when depending on multiple of the
same jar it is hard to know which version will actually be picked

Solution

Default to "pinning" a dependency to the version listed in the BUILD
file.

Result

When directly depending on a jar the version is well known.