Update itertools to 0.8.2 (#8840)
The current version of itertools that the pants engine is using, 0.7, implements a method `flatten()` that now exists in the Rust standard library, so trying to use that method results in a conflict between the two implementations. The maintainers of itertools deprecated their implementation of `flatten` specifically to avoid this conflict, and updating to the latest version picks up this change.