Change signature of store_dict extern (#8844)
### Problem

The `store_dict` extern currently assumes that it will be passed a slice of successive key, value pairs, and checks this at runtime by checking to see if the length of the list is even and panicking if it is not. Runtime panics are bad.

### Solution

Change the extern to instead assume that it will be passed a slice of tuples of key, value pairs, which lets the type system enforce the constraint.