make match() on Enum into a top-level function in meta.py (#8504)
### Problem

Importing via `from pants.util.collections import Enum` induces an indirection which isn't necessary for most users of `Enum` (the ones that don't use the `.match()` function).

### Solution

- Use the normal stdlib enum type.
- Make `match` into a top-level function in `meta.py`.
- Migrate the codebase.

### Result

An awkward import is resolved, and a new import is born! Complete credit for the solution goes to @jsirois in the `#engine` channel on our Slack.