Refactor `option` and `subsystem` tests (#8872)
We'll need to make modifications soon to the `option` folder to unblock progress on transitioning V1 tasks to use V2-style passthrough args. 

Most of the option code is from the early days of Pants. It hasn't had much refactoring and instead has had many new features added on top. This PR takes a step back to try to deduplicate and modernize some of the test code, as prework for these upcoming changes.

A followup will refactor the src code.