Fix isort, Pytest, and MyPy options to be fingerprinted (#8871)
These subsystems recently got new options like `--args` that were accidentally not marked as having `fingerprint=True`. This was missed because in V2 `fingerprint=True` is irrelevant, but now that we've hooked these options up to V1 we need to fix this oversight.

We don't modify `black/subsystem.py` and `flake8/subsystem.py` because those are solely used by V2.