Stop using Java 6 and 7 internally (#8657)
### Problem
Java 6 first came out in 2006 and was ~EOL December 2015. Java 7 came out in July 2011 and was ~EOL July 2019. In 2015, we had to support Java 7 due to a main Pants' user (Square) needing it. Now, these versions are ancient.

Ultimately, we want internal Pants to be using Java 11 (LTS) or possibly Java 13 (non-LTS). This is prework to support that.

### Solution
Use Java 8 everywhere instead of Java 7.

Also, delete some unused testprojects code.