Provide input_files Digest to InteractiveRunner (#8855)
## Problem

When we use InteractiveRunner to run a process not in the workspace but rather in a random tempdir, we need some way to materialize files we might want to work with from a Digest.

## Solution

Add a new field input_files: Digest on InteractiveProcessRequest that works the same way as the analogous field on ExecuteProcessRequest, i.e. provision the working directory of the interactive process with the contents of the Digest.

The code deliberately throws an error rather than trying to write files to disk if we are trying to run the process in the workspace. We already have the Workspace abstraction for writing files there which we can use if we need to do this, and if we force rule-writers to use that abstraction only for touching files in the workspace we can introduce checks to try to mitigate people accidentally overwriting source files they shouldn't be.