Fix `./pants lint2` for Black and isort (#8877)
https://github.com/pantsbuild/pants/pull/8823 broke `./pants lint2` for formatters like Black and Isort because `python_lint_target.py` was not passing the argument `prior_formatter_result_digest` to the constructors for `IsortTarget` and `BlackTarget`.

However, `prior_formatter_result_digest` is only relevant to `fmt2` and the lint logic should never have to deal with the argument.

So, here, we set the field to `Optional` with a default of `None`. `python_fmt_target.py` will ensure that the value is always set when formatting, whereas `python_lint_target.py` will ignore the value as before. This also updates tests to ensure that both code paths are tested.