Deprecate running `python-eval` by default (#8889)
https://github.com/pantsbuild/pants/issues/7444#issuecomment-569821230 proposed deprecating `python-eval` now that Pylint and MyPy provide similar functionality of checking transitive dependencies, e.g. checking for undefined imports. 

These alternatives are much more popular and provide better functionality. At the same time, maintaining `python-eval` adds a burden to the project, especially as we work to migrate the entire Python pipeline to V2.

To be able to remove `python-eval`, we first deprecate using it by default. In a followup, we will deprecate the module entirely.