Use absolute path when running interactive process in tempdir (#8895)
## Problem

When we run an interactive process in a tempdir, we set the current directory of the subprocess to that tempdir with https://doc.rust-lang.org/std/process/struct.Command.html#method.current_dir . According to the linked documentation, in this case it's unpredictable platform-specific behavior whether a relative path will be resolved relative to the directory we set or to the working directory of the parent process.

## Solution

Check and see if the interactive process was passed a relative path, and if so manually canonicalize it to the tempdir. 

While I was at it, I also clarified the text of one of the error messages associated with running interactive processes, to make it clearer where the error was coming from.