Add `deprecated.resolve_conflicting_options` to facilitate moving options (#8903)
Recently, we've moved many options from tasks to subsystems. In the past, we've also moved options from one subsystem to another or renamed options within a subsystem. These use cases are certain to continue, especially in the pursuit of an incremental migration for V2.

We started using a pattern to robustly handle two conflicting options living at the same time during the deprecation. For example, this pattern will provide a friendly error if users try to specify both options at the same time. It will also default to reading from the new option.

However, this pattern was inlined everywhere it was used with the idea that "this is temporary code, so not worth spending too much time deduplicating." This rationale does not stand. While any particular call site may indeed be short-lived, we are going to continue making this type of deprecation in the future. 

Generalizing this functionality ensures that future deprecations are handled robustly and lowers the barrier to entry for deprecating something, thus encouraging an incremental migration for V2.